MB-34017: Correct Disk snapshot marker to not send 0 value
A Disk snapshot will now send a HCS value from the active node to
the replica node. This value could be 0 if no SyncWrites have been
completed. If this is the case, we should not send a 0 HCS as there
is no need and it will cause a pre-condition to fail in the flusher
on the replica.

Change-Id: Ic0708ee4df90d2b1d2fe9f6879fb58838738255f
Reviewed-on: http://review.couchbase.org/113492
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>