MB-35599: Warmup dead vbuckets
Change warmup so that dead and pending vbuckets are warmed-up.
It is very possible that a dead or pending vbucket will next
become replica or active, and as such the data inside it is
absolutely part of the bucket and must not be discarded.

The warmup treats these vbuckets though as replica in terms of
the percentage of quota they may consume.

Change-Id: Ic6ba05b045dfeb5b24ee59001d4dfb51ee135cf5
Reviewed-on: http://review.couchbase.org/113496
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>