BP: MB-35599: Warmup dead vbuckets
Change warmup so that dead and pending vbuckets are warmed-up.
It is very possible that a dead or pending vbucket will next
become replica or active, and as such the data inside it is
absolutely part of the bucket and must not be discarded.

The warmup treats these vbuckets though as replica in terms of
the percentage of quota they may consume.

This is a back port of 8c8a260e

Change-Id: I29ce1a5f3454bf3903acca1065d069a7d5b62c52
Reviewed-on: http://review.couchbase.org/113960
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>