MB-35607: Correct HCS flushing logic
Currently the HCS flushing logic is flawed in two ways:

1) If we disconnect and reconnect a stream then flush a disk
snapshot the HCS may be weakly monotonic (if the active node moves
on and streams to the replica from disk without doing any other
SyncWrites).

2) getItemsForCursor is returning a HCS value before the checkpoint
end. This is incorrect as we can flush partial disk snapshots and a
subsequent warmup may bring the node back in a bad state (missing
prepares). This should work fine if the rest of the disk snapshot is
streamed, but if this node is promoted beforehand and streams from
memory to a replica then the replica will start firing assertions.
This is a dataloss situation already, but undesirable nonetheless.

Change-Id: I700e25d248968ce01abd68236a61fe3a960b11a5
Reviewed-on: http://review.couchbase.org/113552
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>