MB-23086: Do not dlopen any engine shared objects
To enable the thread local storage model of 'initial-exec' in ep-engine
and jemalloc we must no longer dlopen ep.so.

Since memcached is already opening all engines with dlopen during
initialisation we can now move over to just linking the all shared objects
against memcached.

The enginemap code is changed to now call pre-defined symbols of each
engine for the creation of an engine instance and for the shutdown.

engine_testapp now uses BucketType for bucket creation rather than a
path to a shared object. The -E parameter is now changed to select the
BucketType to use for testing and engine_testapp is now linked against
all bucket shared objects.

Change-Id: I3e2f05ee1ee1b3acd75ad9a80c6faa92c8496135
Reviewed-on: http://review.couchbase.org/115504
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>