MB-35654: Make expelling only operate on the oldest checkpoint
Currently, expelling will operate on the oldest checkpoint _containing
cursors_, which may not be the _oldest checkpoint_.

Expelling from a checkpoint which is _not_ the oldest is unsafe as it
leaves holes in the items a cursor will read. This patch prevents this.

If the oldest checkpoint is unref'd (i.e., all cursors are in other
later checkpoints) it can be deleted as a whole; expelling from such a
checkpoint would be inefficient. Given this, trying to expel in this
state should be a noop - the oldest checkpoint will be dropped if it
remains unreferenced, and no other checkpoints are valid candidates.

Change-Id: Id72a613930909103fabd4265d63a5f500567d627
Reviewed-on: http://review.couchbase.org/113706
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>