MB-35331: Remove checkpoint-ID from stored vbucket_state
The checkpoint-id field is persisted so that we can 'crash' and resume
from the current checkpoint-id. However that functionality is to enable
features of the now removed TAP replication protocol, where
active/replicas decided they were in sync by using a checkpoint-ID.

KV still has an incrementing checkpoint-ID which is observable through
stats, but this value is now reset at warm-up.

Change-Id: I480f7577df49a842a46b3804025f0b0a787874d1
Reviewed-on: http://review.couchbase.org/113903
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>