MB-33463: Implement logical delete for magma
Add new kvstoreRev parameter to track kvstore revision. It's
maintained on magmaInfo which is updated as part of vbstate.
Implement prepareToCreateImpl to increment the kvstoreRev.
Implement prepareToDeleteImpl to soft delete the kvstore in magma.
If the vbucket exists, it must call prepareToDelete prior to calling
prepareToCreate or exception will be thrown.

Magma SoftDeleteKVStore effectively deletes the kvstore by moving the
shared pointer to the kvstore to a pending delete list. Subsequent
attempts to access that kvstore in magma will return NotExists status.
The kvstore storage should eventually be deleted when delVBucket is
called. It is at this point we wait for all users of the kvstore to
complete any outstanding operations before removing the storage to the
kvstore.

In the event of an unexpected outage where there are dropped kvstores
still pending, recovery will delete old kvstores up to the latest
kvstore.

Change-Id: Ida4555663c31f79cca53b7ff86270527f49f01ae
Reviewed-on: http://review.couchbase.org/113850
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>