MB-35604: Add ssl_cipher_suite for TLS>=1.3
OpenSSL use a separate API to configure ciphers for TLS 1.3.
For TLS < 1.3 you may use SSL_CTX_set_cipher_list which scans
the list of provided ciphers and selects the ones it supports
and ignores unknown ciphers. The method is considered as
successful if at least one cipher is selected. For TLS 1.3 one
needs to use SSL_CTX_set_ciphersuites and it will _fail_ if
it encounters unknown ciphers.

To work around this ns_server will provide two list of chiphers:

    "ssl_cipher_list" : {
       "tls 1.2" : "list of ciphers to use for TLS1, 1.1 and 1.2",
       "tls 1.3" : "ciphersuites to use for TLS 1.3"
    }

If OpenSSL decides they want to change the API for 1.4 we're ready
for that by adding another field.

This patch also change the logic for what an empty string means.
Previously an empty string indicated that you wanted to use
whatever OpenSSL provided as default. With this patch an empty
string (or not present in the object) means: no ciphers.

Change-Id: If426ed8b02ecb7820851f795e2e2593fa9591a36
Reviewed-on: http://review.couchbase.org/113881
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>