MB-35717: Fix handling of SSL_ERROR_WANT_WRITE
If ssl_write returns SSL_ERROR_WANT_WRITE it has
computed the next segment, but there isn't enough
room in the underlying bio to "send" it so it
needs to be drained.

According to the man page for ssl_write it should
be called again with the same parameters (data and size).

Previously we tried to chunk up our writes to ssl_write
so that they would "fit" into the underlying buffer, but
there isn't much point of doing so (except for adding
extra complexity to an already too complex solution).

Change-Id: I52a0f628f0bb0e5a5e76820b18cec02c467edd78
Reviewed-on: http://review.couchbase.org/113896
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>