MB-32970: getRandomKey to return logical-key when collections disabled
If the client has not enabled collections on their connection then
the returned key must not include the collection prefix.

Note: if developer preview is enabled and other collections exist
a non-collection client issuing getRandomKey will see keys from
those other collections, this is a limitation of the current
getRandomKey operation.

Change-Id: I191a366a6853f0b8b6c2d66e9e7ad2a7ad147b31
Reviewed-on: http://review.couchbase.org/113942
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>