MB-35499: Only seqnoAck over an fully established stream
Only seqnoAck as a replica after we have had acknowledgement of the
stream by the producer for a given vbucket, otherwise the producer may
not have context for why the seqnoAck is being sent and will log
an error that it does not have a StreamContainer for the vbucket. This
will also close the connection with the replica, due to the exception
being thrown in the packet execution runloop.

Update dcp_stream_test.cc to correctly update the DCP
stream's state so that we enable the ability to seqnoAck when testing
when seqnoAcks should be sent and add helper methods to perform
state transitions.

Also remove transition from StreamState::Reading to
StreamState::Pending as being valid for a PassiveStream as we never
do this and allows us to treat the Pending state as an unacknowledged
stream by the producer.

Change-Id: I67d43fa8d1cf2bbf3b7c0c14febf4ab8a21925d7
Reviewed-on: http://review.couchbase.org/113946
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>