MB-35707: Don't dereference null chainPtr during ActiveDM::commit
If debug logging is enabled then ActiveDM::commit() prints out each
node which ack'd the SyncWrite. However, since
5191f1ceb4946c4bf1d7a166e25d3ba1e2783279 the chainPtrs are set to null
when a SyncWrite is resolved and moved to the ResolvedQueue (as they
could point to Chains which are no longer valid).

As such, it is invalid to dereference the chainPtrs in commit() now.

Given this code was only used for debug log messages (and isn't
particulary useful), simply remove the offending code.

Change-Id: Ifbad854d30dce45622f875ecbe45c02f5ee120b7
Reviewed-on: http://review.couchbase.org/113943
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>