MB-35633: Store unprepared aborts from disk snapshots
Prior to this patch, aborts received as part of a disk snapshot where
an associated prepare has /not/ been received would be allowed (because
aborts can dedupe the prepare) but would be ignored.

This was first noticed as a problem during unrelated testing in which
closing a checkpoint after a snapshot could try to close an /empty/
checkpoint, if the only item the snapshot contained was an abort which
had deduped the matching prepare.

More significantly, ignoring aborts if the prepare was unseen can lead
to a different replica holding a prepare that will never receive a
completion. Consider one replica connected and received only the abort,
but a second replica received the prepare before it was deduped, but
has not /yet/ received the abort:

Active    : Pre Abort
Replica 1 : xxx Abort
Replica 2 : Pre

If the active fails and replica 1 is promoted:

 - R1 knows nothing of the prepare, and discarded the abort.
 - R2 has the prepare
 - R2 connects to the active, but the abort will never be sent

This patch avoids this by storing aborts for persistence and
replication even if the matching prepare has been deduped.

Change-Id: Iaadf1b459cdb4cdb14a56579be6170814fb4e1ed
Reviewed-on: http://review.couchbase.org/113963
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>