MB-35741 [SR]: Decrement onDiskPrepares during compaction
During compaction, if a Prepared SyncWrite is discarded (compacted
away), the on-disk prepare count in vbState is not updated. This later
becomes a problem during a subsequent warmup, when a vBuckets' item
count is re-calculated - we can end up with a negative item count.

This manifests the next time the numTotalItems is updated - typically
the next time the flusher runs, resulting in the following exception
being thrown:

    Caught unhandled std::exception-derived exception. what(): ThrowExceptionUnderflowPolicy

Fix by ensuring that every time a prepare is discarded we increment
CompactionStats::preparesPurged, then apply that delta to vbstate at
the end of compaction.

Change-Id: I3be0ecfe119db8b42e9246034cc67cd8cb79323f
Reviewed-on: http://review.couchbase.org/113964
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>