MB-33463: Change calls to SyncCommitBatches to use config param
SyncCommitBatches is used to persist any pending batches of items to the
WAL. In addition, it can also create a commit point when configuration
calls for creating a commit point on every batch flush. This checkin is
to make all calls to SyncCommitBatch use the configuration setting.

This issue was exposed during testing when it was observed that
kv_engine is calling snapshotVBucket quite frequently, especially in low
DGM scenarios forcing the write cache out and thus creating tiny
sstables which increased compaction frequency.

Change-Id: I06b2e6adf3ae2ea3653ddb6a77f4afd5c4a79104
Reviewed-on: http://review.couchbase.org/114095
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>