MB-35150 Ensure GetReplica returns only committed values
Ensure GetReplica returns committed values instead of
ENGINE_SYNC_WRITE_RECOMMIT_IN_PROGRESS when a new item is in the state
PreparedMaybeVisible for the key specified by the GetReplica op.

To ensure we do this, VBucket::getInternal() must be informed it is
called for a GetReplica. If so we want to make sure we only look for
committed values in the hash table and ensure that the value we find is
not in the state CommittedState::PreparedMaybeVisible. To do this we
need to pass the context that we are performing a replica op to
fetchValidValue(). Which allows us not to return any item in the
PreparedMaybeVisible state.

Also refactor KVBucketIface::getInternal() interface and
implementations so that they take an enum to inform the method if the
get is for a replica vbucket or not. As opposed to the allowedState
parameter which is only ever in the vbucket_state_active or
vbucket_state_replica state.

Change-Id: I11fe1fecdf66ee5c0abad3119becc5039eb96375
Reviewed-on: http://review.couchbase.org/114525
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>