MB-35764: Do not try to close checkpoint on snapEnd
Prior to this patch on receiving a snapshotEnd, a PassiveStream can
attempt to close the open checkpoint if mem_used is above a threshold.
The original intention of this was to try to free memory tied up in the
checkpoint by closing it and allowing it to be deleted when unreffed.

Recently checkpoint expelling was added which can recover memory from
a checkpoint by freeing individual items without requiring the entire
checkpoint to be deleted.

With expelling, closing a checkpoint in this situation is no longer
necessary and is removed with this patch.

Now checkpoints on a replica will only be closed when instructed by the
active (snapshot with the CHK flag).

Change-Id: I30a24a9293b4b63c4940514ec20c792751eaba3c
Reviewed-on: http://review.couchbase.org/114117
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>