MB-35768 [SR]: Allow warmup of SyncWrites with impossible topology
If warmup occurs with:

1) One or more prepared SyncWrites, and
2) A topology which durability isn't possible for (there isn't a
   majority of nodes present)

Then warmup currently crashes as the following Expect inside
checkDurabilityPossibleAndResetTopology() fails when constructing
SyncWrites from warmup:

    Expects(firstChain.isDurabilityPossible());

In general this is a valid check - we should only accept "new"
SyncWrites from clients if there's a topology which durabilty is
possible (otherwise the client should have received
EDurabilityImpossible).

However in the case where the SyncWrite was already accepted (before
the node restarted and warmed back up), then we *must* keep the
SyncWrite in the ActiveDM.

Therefore remove the isDurabilityPossible check when constructing
SyncWrites with an infinite timeout (i.e. from Warmup or PassiveDM).

Note: PrePostStateChecker changes were necessary to avoid the
EXPECT_TRUE(vb) check firing during the destruction of
PrePostStateChecker when the gsl::fail_fast exception was thrown prior
to the fix. With the change to PrePostStateChecker there's a clean
failure in the regression test before the fix is applied.

Change-Id: Iaeb405efa3866f3481dd0778de75b2f2c81768c4
Reviewed-on: http://review.couchbase.org/114137
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>