MB-35631: Only persist HCS after complete disk snapshot persisted
Currently, the HCS sent as part of a disk snapshot marker will be
persisted the first time any items from the checkpoint are flushed to
disk. This lead to non-monotonicity as the flusher would later try to
persist a /lower/ HCS upon encountering a commit.

This patch ensures the disk snapshot HCS will only be persisted if the
current flush batch reaches the end of the disk checkpoint.

Change-Id: I4fc6326c6c4db814a43ebe6f552d9d3bc96b5448
Reviewed-on: http://review.couchbase.org/114245
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>