MB-35812: Ensure correct cursor is selected when expelling from ckpt
Checkpoint expelling expects that the "earliest" cursor will reside in
the oldest referenced checkpoint. While this is correct, "earliest" was
previously defined as the cursor pointing to the item with the lowest
seqno.

However, the same seqno can in certain situations appear in two
checkpoints (if a checkpoint ends with a set vb state, see
CheckpointManager::createCheckpointItem for the logic which sets the
seqno for the meta item). Given this, the ordering used to determine
the earliest cursor is insufficient, and could select the later cursor.

To resolve this, find the cursor with the lowest seqno and
checkpointId.

Change-Id: Ie28279a4a21fb6e9804bca0f370950ed135e97cb
Reviewed-on: http://review.couchbase.org/114376
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>