[cluster_testapp] Let the bufferevent close its sockets
We had an explicit close of the socket and used unique_ptrs for
the bufferevent structure. Calling bufferevent_free will remove
the listening mode for the socket, but if the socket is closed
we're trying to operate on an "invalid" socket. Given that
there is an option to bufferevent to close it when its done
with it lets use that so that we don't have to care about
the ordering

Change-Id: I36f29f04163e2024590e34418a177c034fb5bff4
Reviewed-on: http://review.couchbase.org/114370
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>