MB-35592: Fix hang situation in cluster_test
The test for MB34780 tries to shut down the replicator
when it sees a DCP_PREPARE being transmitted over the
DCP stream. The problem was however that it started
off by deleting the default bucket and then recreate
the default bucket _WITHOUT_ deleting the database
files first. That means that the second time it
created the database it would actually "recover" the
old database and a prepare could be resent (and
depending on the timing we might not get all of the
dcp streams set up before we started to drop packets)

Change-Id: Ia8719618655330876c2430bc42ca9cd6164a04fd
Reviewed-on: http://review.couchbase.org/114439
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>