MB-31450: Track max-deleted revision using checkpoint::queueDirty
When a checkpoint encounters a DEL(k1) then SET(k1) it will discard the
DEL(k1) (de-duplication). A consequence of this is that the flusher will
not be able to update the vbucket meta-data (vbstate) with the correct
"max-deleted revision seqno".

As shown in the test this could mean that conflict resolution could
incorrectly allow a mutation because a warmup occurred.

To address this issue, EPBucket::flushVBucket no longer inspects the
items looking for deletes, instead Checkpoint:;queueDirty does the
inspection before de-duplication occurs. Each Checkpoint now has an
optional member which when initialised, stores the largest revision
seqno of all queued deletes.

The flusher now gets the value to persist from the checkpoint-manager.

Change-Id: I5dfa7a4cb1e0359eef83b7e1df84f9448c57c43c
Reviewed-on: http://review.couchbase.org/114442
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>