MB-35932: Distinguish a bg-fetch needed vs pending durability
The KVBucket set/add/replace/deleteItem methods were overloading
the meaning of EWOULDBLOCK for pending sync-writes, in EWOULDBLOCK
was returned for sync-write accepted and bg-fetch needed. The
caller of the method then assumed a pending set for example
returning EWOULDBLOCK was always a sync-write accepted, breaking
SET with CAS in full-eviction (and others).

To address this issue, add a new error code ENGINE_SYNC_WRITE_PENDING
so that the caller of KVBucket methods can distinguish sync-write
accepted vs needs bg-fetch.

Change-Id: I91cc4883dc15c145ff9293ad680beb9063028e66
Reviewed-on: http://review.couchbase.org/114596
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>