MB-35933: Ensure duplicate prepare replaces correct prepare
In cases where a prepare is allowed to replace an older prepare (after
a disconnect/reconnect, if the intervening commit/abort may have been
deduped), ensure completed prepares are not considered for
replacement.

Completed prepares may not be immediately removed from tracked writes
if they are PersistToMajority, as they must be kept until the prepare
is persisted.

A duplicating prepare will find the existing prepare in the hashtable,
and is only permitted to replace *that* prepare in trackedWrites. Prior
to this change, an `Expects` would fail if an older completed prepare
is encountered first.

This patch ignores completed prepares when searching for the correct
SyncWrite to replace.

Change-Id: I85bc660441a6036a53218b5f42a53229bc6dc8ae
Reviewed-on: http://review.couchbase.org/114615
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>