MB-35934: Skip deleted vBuckets in DurabilityCompletionTask
The task incorrectly assumed that if it was scheduled for a given
vBucket then the vBucket would still exist. Skip these vBuckets if
we cannot acquire the shared_ptr to them.

Also, fix an issue where we incorrectly did not wake the task on the
first SyncWrite that should schedule a vBucket (sets pendingVBs[vbid]
from false to true). Previously we would only wake the task for
any other transition (e.g. true to true).

Change-Id: I176e68b9f605408b5f704023c30847ad56ebd79d
Reviewed-on: http://review.couchbase.org/114619
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>