MB-35954: Ensure durable sets with cas compare to the committed cas
When storing with cas, currently the cas is compared against the cas of
the stored value which is to be modified, whether it is a completed
prepare or a committed item.

In persistent buckets, the stored value to be modified is always the
previous committed item, as prepares are not kept in the hashtable
after completion so this is not an issue.

However, in ephemeral buckets, completed prepare stored values remain
in the hashtable to be modified by a new durable write, but are not
necessarily "in-sync" with the committed value if non-durable writes
are present.

Because of this, sets need to test the cas against that of the
committed stored value _even if_ they are about to modify a completed
prepare stored value.

Change-Id: I6016d9e4ae373ab8daefd92bc291e02602bb01fb
Reviewed-on: http://review.couchbase.org/114812
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>